+++ library/Kathu/Graphics/Camera.hs
@@ -4,4 +4,4 @@ module Kathu.Graphics.Camera where
 
 import GHC.Generics
 
-newtype Camera = Camera {zoom :: Float} deriving (Show, Eq, Generic)
+newtype Camera = Camera {zoom :: Double} deriving (Show, Eq, Generic)
